server/requirements.txt
openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git@v0.2.3
fastapi>=0.115.0
pydantic>=2.0.0
uvicorn>=0.24.0
requests>=2.31.0
